changed model

+++ models/goals.go
@@ -9,7 +9,7 @@ type Goals struct {
 	BookId       string    `json:"bookId" db:"book_id"`
 	CreatedAt    time.Time `json:"createdAt" db:"created_at"`
 	EndDate      time.Time `json:"endDate" db:"end_date"`
-	GoalWords    int       `json:"goalWords" db:"goal_words"`
+	GoalWords    float64   `json:"goalWords" db:"goal_words"`
 	Id           uuid.UUID `json:"id" db:"id"`
 	IsFinished   bool      `json:"isFinished" db:"is_finished"`
 	StartDate    time.Time `json:"startDate" db:"start_date"`
@@ -17,7 +17,7 @@ type Goals struct {
 	UpdatedAt    time.Time `json:"updatedAt" db:"updated_at"`
 	UserId       string    `json:"-" db:"user_id"`
 	Description  *string   `json:"description" db:"description"`
-	WrittenWords int       `json:"writtenWords" db:"written_words"`
+	WrittenWords float64   `json:"writtenWords" db:"written_words"`
 	WordsPerDay  float64   `json:"wordsPerDay" db:"words_per_day"`
 	IsExpired    bool      `json:"isExpired" db:"is_expired"`
 }
