+++ models/reports.go
@@ -6,7 +6,7 @@ type ReportsModel struct {
 	GoalId      string  `json:"goalId" db:"goal_id"`
 	Id          string  `json:"id" db:"id"`
 	UpdatedAt   string  `json:"updatedAt" db:"updated_at"`
-	WordsAmount int     `json:"wordsAmount" db:"words_amount"`
+	WordsAmount float64 `json:"wordsAmount" db:"words_amount"`
 	Description *string `json:"description" db:"description"`
 	Title       string  `json:"title" db:"title"`
 	UserId      string  `json:"-" db:"user_id"`
